+++ tests/test_admin.py
@@ -1,3 +1,4 @@
+import django
 import django.contrib.admin.utils as admin_utils
 
 import pytest
@@ -22,7 +23,9 @@ INTEGER_FIELD = ModelWithVanillaMoneyField._meta.get_field("integer")
     ),
 )
 def test_display_for_field_with_legacy_formatting(legacy_formatting, settings, value, expected):
-    settings.USE_L10N = True
+    # This now defaults to True and raises RemovedInDjango50Warning
+    if django.VERSION < (4, 0):
+        settings.USE_L10N = True
     # This locale has no definitions in py-moneyed, so it will work for localized money representation.
     settings.LANGUAGE_CODE = "cs"
     settings.DECIMAL_PLACES_DISPLAY = {}
@@ -40,7 +43,9 @@ def test_display_for_field_with_legacy_formatting(legacy_formatting, settings, v
     ),
 )
 def test_display_for_field(settings, value, expected):
-    settings.USE_L10N = True
+    # This now defaults to True and raises RemovedInDjango50Warning
+    if django.VERSION < (4, 0):
+        settings.USE_L10N = True
     # This locale has no definitions in py-moneyed, so it will work for localized money representation.
     settings.LANGUAGE_CODE = "cs"
     assert admin_utils.display_for_field(value, MONEY_FIELD, "") == expected