+++ tests/test_form.py
@@ -190,7 +190,7 @@ class TestDisabledField:
 def test_decimal_places_model_form(model_class):
     """Forms should use DECIMAL_PLACES setting value when none specified."""
 
-    expected = str(10 ** -settings.DECIMAL_PLACES)
+    expected = str(10**-settings.DECIMAL_PLACES)
     assert model_class().fields["money"].widget.widgets[0].attrs["step"] == expected
 
 
@@ -198,5 +198,5 @@ def test_precedence_decimal_places_model_form():
     """Forms should use decimal_places in field value when specified."""
 
     decimal_places = PreciseModelForm.Meta.model._meta.fields[2].decimal_places
-    expected = str(10 ** -decimal_places)
+    expected = str(10**-decimal_places)
     assert PreciseModelForm().fields["money"].widget.widgets[0].attrs["step"] == expected