+++ tests/test_models.py
@@ -5,6 +5,7 @@ Created on May 7, 2011
 """
 import datetime
 from copy import copy
+from decimal import Decimal, InvalidOperation
 
 from django import VERSION
 from django.core.exceptions import ValidationError
@@ -18,6 +19,7 @@ import pytest
 from djmoney.models.fields import MoneyField
 from djmoney.money import Money
 from moneyed import Money as OldMoney
+from moneyed.classes import CurrencyDoesNotExist
 
 from .testapp.models import (
     AbstractModel,
@@ -40,6 +42,7 @@ from .testapp.models import (
     ModelWithDefaultAsStringWithCurrency,
     ModelWithNonMoneyField,
     ModelWithNullableCurrency,
+    ModelWithNullDefaultOnNonNullableField,
     ModelWithSharedCurrency,
     ModelWithTwoMoneyFields,
     ModelWithUniqueIdAndCurrency,
@@ -301,6 +304,80 @@ class TestVanillaMoneyField:
         instance = NullMoneyFieldModel.objects.create()
         assert instance.field is None
 
+    def test_implicit_currency_field_not_nullable_when_money_field_not_nullable(self):
+        instance = ModelWithVanillaMoneyField(money=10, money_currency=None)
+        with pytest.raises(TypeError, match=r"Currency code can't be None"):
+            instance.save()
+
+
+@pytest.mark.parametrize(
+    ("default", "error", "error_match"),
+    [
+        pytest.param("10", AssertionError, r"Default currency can not be `None`", id="string without currency"),
+        pytest.param(b"10", AssertionError, r"Default currency can not be `None`", id="bytes without currency"),
+        pytest.param(13.37, AssertionError, r"Default currency can not be `None`", id="float"),
+        pytest.param(Decimal(10), AssertionError, r"Default currency can not be `None`", id="decimal"),
+        pytest.param(10, AssertionError, r"Default currency can not be `None`", id="int"),
+        pytest.param("100 ", CurrencyDoesNotExist, r"code  ", id="string with trailing space and without currency"),
+        pytest.param(b"100 ", CurrencyDoesNotExist, r"code  ", id="bytes with trailing space and without currency"),
+        pytest.param("100 ABC123", CurrencyDoesNotExist, r"code ABC123", id="string with unknown currency code"),
+        pytest.param(b"100 ABC123", CurrencyDoesNotExist, r"code ABC123", id="bytes with unknown currency code"),
+        # TODO: Better error reporting on > 1 white spaces between amount and currency as that could be
+        # quite difficult to spot?
+        pytest.param(
+            "100  SEK", CurrencyDoesNotExist, r"code  SEK", id="string with too much value and currency separation"
+        ),
+        pytest.param(
+            b"100  SEK", CurrencyDoesNotExist, r"code  SEK", id="bytes with too much value and currency separation"
+        ),
+        pytest.param(
+            "  10 SEK  ",
+            InvalidOperation,
+            r"(ConversionSyntax|Invalid literal)",
+            id="string with leading and trailing spaces",
+        ),
+        pytest.param(
+            b"  10 SEK  ",
+            InvalidOperation,
+            r"(ConversionSyntax|Invalid literal)",
+            id="bytes with leading and trailing spaces",
+        ),
+        pytest.param(
+            "  10 SEK", InvalidOperation, r"(ConversionSyntax|Invalid literal)", id="string with leading spaces"
+        ),
+        pytest.param(
+            b"  10 SEK", InvalidOperation, r"(ConversionSyntax|Invalid literal)", id="bytes with leading spaces"
+        ),
+        pytest.param("10 SEK  ", CurrencyDoesNotExist, r"code SEK   ", id="string with trailing spaces"),
+        pytest.param(b"10 SEK  ", CurrencyDoesNotExist, r"code SEK   ", id="bytes with trailing spaces"),
+    ],
+)
+def test_errors_instantiating_money_field_with_no_default_currency_and_default_as(default, error, error_match):
+    with pytest.raises(error, match=error_match):
+        MoneyField(max_digits=10, decimal_places=2, default=default, default_currency=None)
+
+
+@pytest.mark.parametrize(
+    ("default", "default_currency", "expected"),
+    [
+        pytest.param("10 SEK", None, Money(10, "SEK"), id="string with currency"),
+        pytest.param(b"10 SEK", None, Money(10, "SEK"), id="bytes with currency"),
+        pytest.param("10", "USD", Money(10, "USD"), id="string without currency and currency default"),
+        pytest.param(b"10", "USD", Money(10, "USD"), id="bytes without currency and currency default"),
+    ],
+)
+def test_can_instantiate_money_field_default_as(default, default_currency, expected):
+    field = MoneyField(max_digits=10, decimal_places=2, default=default, default_currency=default_currency)
+    assert field.default == expected
+
+
+def test_errors_on_default_values_being_none_when_fields_have_not_null_constraint():
+    instance = ModelWithNullDefaultOnNonNullableField()
+    assert instance.money is None
+    assert instance.money_currency is None
+    with pytest.raises(IntegrityError, match=r"testapp_modelwithnulldefaultonnonnullablefield.money"):
+        instance.save()
+
 
 class TestGetOrCreate:
     @pytest.mark.parametrize(
@@ -370,6 +447,12 @@ class TestNullableCurrency:
     def test_fails_with_null_currency(self):
         with pytest.raises(ValueError) as exc:
             ModelWithNullableCurrency.objects.create(money=10)
+        assert str(exc.value) == "Missing currency value"
+        assert not ModelWithNullableCurrency.objects.exists()
+
+    def test_fails_with_null_currency_decimal(self):
+        with pytest.raises(ValueError) as exc:
+            ModelWithNullableCurrency.objects.create(money=Decimal(10))
         assert str(exc.value) == "Missing currency value"
         assert not ModelWithNullableCurrency.objects.exists()
 
@@ -712,10 +795,12 @@ def test_override_decorator():
 def test_properties_access():
     with pytest.raises(TypeError) as exc:
         ModelWithVanillaMoneyField(money=Money(1, "USD"), bla=1)
-    if VERSION[:2] > (2, 1):
+    if VERSION[:2] > (4, 0):
+        assert str(exc.value) == "ModelWithVanillaMoneyField() got unexpected keyword arguments: 'bla'"
+    elif VERSION[:2] > (2, 1):
         assert str(exc.value) == "ModelWithVanillaMoneyField() got an unexpected keyword argument 'bla'"
     else:
-        assert str(exc.value) == "'bla' is an invalid keyword argument for this function"
+        assert str(exc.value) == "ModelWithVanillaMoneyField() got an unexpected keyword argument 'bla'"
 
 
 def parametrize_with_q(**kwargs):
@@ -787,3 +872,34 @@ def test_mixer_blend():
         instance = mixer.blend(ModelWithTwoMoneyFields)
         assert isinstance(instance.amount1, Money)
         assert isinstance(instance.amount2, Money)
+
+
+@pytest.mark.parametrize(
+    ("attribute", "build_kwargs", "expected"),
+    [
+        pytest.param(
+            "default_currency",
+            {"max_digits": 9, "null": True, "default_currency": None},
+            None,
+            id="default_currency_as_none",
+        ),
+        pytest.param(
+            "default_currency",
+            {"max_digits": 9, "null": True, "default_currency": "SEK"},
+            "SEK",
+            id="default_currency_as_non_default_not_none",
+        ),
+        pytest.param(
+            "currency_max_length",
+            {"max_digits": 9, "currency_max_length": 4},
+            4,
+            id="currency_max_length_as_non_default",
+        ),
+    ],
+)
+def test_deconstruct_includes(attribute, build_kwargs, expected):
+    instance = MoneyField(**build_kwargs)
+    __, ___, args, kwargs = instance.deconstruct()
+    new = MoneyField(*args, **kwargs)
+    assert getattr(new, attribute) == getattr(instance, attribute)
+    assert getattr(new, attribute) == expected